+++ R/read-file.R
@@ -23,6 +23,7 @@
 #' rdata <- read_file(file = "/file/path/mydata.xlsx", sheet = "S3", range = cell_rows(1:4))
 #' rdata <- read_file(file = "/file/path/mydata.csv", sep = ",", header = FALSE)
 #' }
+#' @importFrom methods is
 #' @export
 read_file <- function(file = NULL, ...) {
   is_debug()
@@ -37,7 +38,7 @@ read_file <- function(file = NULL, ...) {
 
   dt <- find_data(file, ...)
 
-  if (sum(is.element(class(dt), "data.table")) == 0){
+  if (isFALSE(is(dt, "data.table"))){
     data.table::setDT(dt)
   }