+++ inst/recode-aggregate.sql
@@ -1,0 +1,3 @@
+select FILGRUPPE, KOL, TYPE, FRA, TIL
+from tbl_Kode
+where FILGRUPPE = '%s' and VERSJONTIL = #9999-01-01#